postgres/requirements.txt
asgiref==3.3.4
click==8.0.1
decouple==0.0.7
fastapi==0.65.1
h11==0.12.0
psycopg2-binary==2.8.6
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
